Added comments, renamed variables, added environment
