+++ .gitignore
@@ -105,8 +105,10 @@ celerybeat.pid
 # Environments
 .env
 .venv
+.arcvenv
 env/
 venv/
+arcvenv/
 ENV/
 env.bak/
 venv.bak/
